renamed config.cpp to config.cc and moved the config lib to a better folder
